Fix browser freeze on stock page with large holdings from pre-split candle prices
- Normalize historical candle OHLC prices by cumulative split ratio before passing to
  lightweight-charts, preventing freezes caused by pre-split prices exceeding CHART_SAFE_MAX (1e9)
- Fix OrderBookDto.OrderBookEntry.quantity: long → BigDecimal to handle post-split
  order quantities that may exceed Long.MAX_VALUE
- Fix StockService.toOrderBookEntries: replace longValue() cast with BigDecimal conversion

+++ backend/src/main/java/com/spotchzxk/presentation/dto/OrderBookDto.java
@@ -17,7 +17,7 @@ public class OrderBookDto {
     @AllArgsConstructor
     public static class OrderBookEntry {
         private BigDecimal price;
-        private long quantity;
+        private BigDecimal quantity;
     }
 }
